backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.3.0
opencv-python==4.8.1.78
deepface==0.0.79
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
tensorflow==2.15.0
